Finished Writing EMG-IMU Upsampler
Improved the EMG-IMU Upsampler to use the same process as within the Combine Upsampler.

As well as Improving the EMG-IMU Upsampler UI to show the current file being upsampled.

There's an Issue with the end of files that needs fixed. The end of files end with empty (Zero) data that needs to be removed